pyproject.toml
[project]
name = "diffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.3.1",
    "torchvision==0.18.1",
    "numpy>=2.1.0",
    "opencv-python>=4.10.0.84",
    "plotly>=5.23.0",
    "streamlit>=1.37.1",
    "ipython>=8.26.0",
    "lightning>=2.4.0",
    "rootutils>=1.0.7",
    "rich>=13.7.1",
    "hydra-core>=1.3.2",
    "hydra-colorlog>=1.2.0",
    "einops>=0.8.0",
    "accelerate>=0.31.0",
    "ema-pytorch>=0.6.2",
    "scipy>=1.14.1",
    "aim>=3.24.0",
    "matplotlib>=3.9.2",
    "mlflow>=2.16.0",
    "wandb>=0.17.8",
    "tensorboard>=2.17.1",
    "imageio>=2.36.0",
    "torch-fidelity>=0.3.0",
    "fire>=0.7.0",
    "datasets>=4.0.0",
    "pyarrow>=17.0.0",
]

[project.scripts]
train = "diffusion.scripts.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
    "pytest>=8.3.2",
    "ipython>=8.26.0",
    "viztracer>=0.17.1",
    "watchdog>=4.0.2",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
# ignore = ["C901", "E501", "E741", "F402", "F823" ]
select = ["C", "E", "F", "I", "N"]
ignore = ["N812"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 20


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as fast unit tests",
    "smoke: marks tests as smoke tests",
    "network: marks tests that require network access",
]
